Fix busy time config
Fix per Marios's comment:
this is the data-bundle :

"endpoint": "calendar/google/events",
                        "mapping": {
                            "created": "created"
                        }

That tells the HAT that from calendar/google/events the function only requires the created field. In this case HAT will only return the created fields for all the records.on busy_time.py in line 39:

        if 'start' in record and 'end' in record and 'status' in record:

The function checks for start end and status fields. HAT returned only the created fields as it was asked in the data debit. I didn’t check the whole file for what other values busy-time requires.